src/server/python-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
snowflake-connector-python==3.6.0
requests==2.31.0